Towers
TLDR;
* Targeting styles:
  * FAR: targets the farthest enemy
  * CLOSE: targets the CLOSEST ENEMY TO THE TOWER
  * CASTLE: targets the enemy CLOSEST TO THE CASTLE
  * WEAK: UNIMPLEMENTED but supports different enemy types when they get implemented.
  * STRONG: UNIMPLEMENTED but supports different enemy types when they get implemented.

* Projectiles aim and fly towards the targeted enemy and won't destroy or get stuck on misc objects in scene.

Implemented the Basic functionality of the towers. they will target enemies with the layer "Enemy" with given projectile prefabs. there is no actual damage or destroying of objects because there is no enemy framework to build off. the framework for towers is built in abstract classes that build off each other so you can create new types or rework the current one to work with any "theme" you want.

Check the Towers Scene to see them in action.